Await responses before loading stuff. Filters ahead

+++ app/src/main/java/com/example/bookstoreapp/retrofit/ApiService.kt
@@ -64,7 +64,7 @@ interface ApiService {
     @Headers("ngrok-skip-browser-warning: 69420")
     @DELETE("categories/{id}")
     suspend fun deleteCategory(
-        @Path("id") id: Int,
+        @Path("id") id: Short,
         @Header("Token") idToken: String
     )
 }
